+++ frontend/src/view/campaign/create-edit/blocs/annotators.bloc.tsx
@@ -1,9 +1,9 @@
 import React, { ChangeEvent, useEffect, useMemo, useState } from "react";
 import { createCampaignActions } from "@/slices/create-campaign";
-import { useUsersAPI } from '@/services/api/user';
+import { useUsersAPI } from '@/services/api/user.ts';
 import { FormBloc, Searchbar, ChipsInput, Input } from "@/components/form";
 import { Item } from "@/types/item";
-import { User } from '@/types/user.ts';
+import { User } from '@/types/user';
 import { useAppSelector, useAppDispatch } from "@/slices/app";
 import { useToast } from "@/services/utils/toast";
