+++ frontend/src/view/campaign/create-edit/blocs/import-modal/detectors-content.tsx
