+++ frontend/src/view/campaign/create-edit/create-campaign.page.tsx
@@ -2,8 +2,8 @@ import React, { FormEvent, useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { IonButton, IonSpinner } from "@ionic/react";
 import { useCreateCampaign } from "@/services/create-campaign";
-import { useBlur } from "@/services/utils/clic";
-import { useToast } from "@/services/utils/toast";
+import { useBlur } from "@/services/utils/clic.ts";
+import { useToast } from "@/services/utils/toast.ts";
 import { GlobalInfoBloc } from "./blocs/global-info.bloc.tsx";
 import { DatasetBloc } from "./blocs/dataset.bloc.tsx";
 import { AnnotatorsBloc } from "./blocs/annotators.bloc.tsx";
