+++ frontend/src/view/campaign/detail/annotation-campaign-detail.page.tsx
