+++ frontend/src/view/campaign/list/annotation-campaign-list.page.css
