+++ frontend/src/view/campaign/list/campaign-card/campaign-card.component.tsx
